added more pop ups

+++ Views/FileOverviewPage.xaml.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 using FileManager.Models;
 using FileManager.ViewModels;
 using SharpHook;
@@ -177,7 +178,7 @@ public partial class FileOverviewPage : ContentPage
         }
     }
 
-    private void RightContextClick(object sender, EventArgs e)
+    private async void RightContextClick(object sender, EventArgs e)
     {
         MenuFlyoutItem item = (MenuFlyoutItem)sender;
         if (item.Text == "Refresh")
@@ -191,13 +192,17 @@ public partial class FileOverviewPage : ContentPage
         }
         else if (item.Text == "Delete")
         {
-            viewModel.RightSideViewModel.DeleteItem();
-            System.Diagnostics.Debug.WriteLine("HALLOOOO? R");
+            if (LeftCollection.SelectedItems.Count == 0)
+            {
+                await DisplayAlert("Alert", "You have to select first to delete", "OK");
+                return;
+            }
 
+            viewModel.RightSideViewModel.DeleteItem();
         }
     }
 
-    private void LeftContextClick(object sender, EventArgs e)
+    private async void LeftContextClick(object sender, EventArgs e)
     {
         MenuFlyoutItem item = (MenuFlyoutItem)sender;
         if (item.Text == "Refresh")
@@ -211,9 +216,13 @@ public partial class FileOverviewPage : ContentPage
         }
         else if (item.Text == "Delete")
         {
-            viewModel.LeftSideViewModel.DeleteItem();
-            System.Diagnostics.Debug.WriteLine("HALLOOOO?");
+            if(LeftCollection.SelectedItems.Count == 0)
+            {
+                await DisplayAlert("Alert", "You have to select first to delete", "OK");
+                return; 
+            }
 
+            viewModel.LeftSideViewModel.DeleteItem();
         }
     }
